CPS_public/backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.23.2
sqlalchemy==2.0.21
pydantic[email]==1.10.13
python-dotenv==1.0.0
celery==5.3.4
redis==5.0.1
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
playwright==1.39.0
requests==2.31.0
psycopg2-binary==2.9.9
alembic==1.12.0
starlette==0.27.0
python-decouple==3.8